+++ src/test/java/kitchenpos/apply/menus/tobe/infra/FakePurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.apply.menu.tobe.infra;
+package kitchenpos.apply.menus.tobe.infra;
